requirements.txt
torch==2.3.1+cu121
pillow==10.2.0
opencv-python==4.10.0.84
safetensors==0.4.3
onnx==1.16.1
onnxruntime==1.18.1
